Clean up code.
