+++ plugin/src/main/java/io/spine/tools/bootstrap/CodeGenExtension.java
@@ -22,7 +22,6 @@ package io.spine.tools.bootstrap;
 
 import com.google.errorprone.annotations.OverridingMethodsMustInvokeSuper;
 import io.spine.logging.Logging;
-import io.spine.tools.bootstrap.ProtobufGenerator.ProtocPlugin;
 import org.gradle.api.Project;
 
 import static io.spine.tools.bootstrap.SpineModule.base;
